Removed check for movetype in seismic sensor and made seismicsignature default to 0 for non ground units
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1679 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Units/Unit.h
@@ -166,7 +166,7 @@ public:
 	int jammerRadius;
 	int sonarJamRadius;
 	int seismicRadius;
-	int seismicSignature;
+	float seismicSignature;
 	bool hasRadarCapacity;
 	std::vector<int> radarSquares;
 	int2 oldRadarPos;
